enforce numpy type

+++ pysrc/test_libnpy.py
@@ -2,7 +2,7 @@ import numpy as np
 
 def gen_npy(save_str):
 
-    data = np.array([3.0,4.0,5.0,6.0])
+    data = np.array([3.0,4.0,5.0,6.0], dtype=np.float32)
 
     np.save(save_str, data)
